+++ common/src/main/java/com/lowdragmc/lowdraglib/utils/DummyWorld.java
@@ -2,11 +2,11 @@ package com.lowdragmc.lowdraglib.utils;
 
 import com.google.common.base.Suppliers;
 import com.lowdragmc.lowdraglib.LDLib;
-import com.lowdragmc.lowdraglib.Platform;
 import com.lowdragmc.lowdraglib.client.scene.ParticleManager;
 import com.lowdragmc.lowdraglib.core.mixins.accessor.ParticleEngineAccessor;
 import com.lowdragmc.lowdraglib.utils.virtual.WrappedClientWorld;
 import lombok.Getter;
+import lombok.Setter;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.MethodsReturnNonnullByDefault;
@@ -14,20 +14,17 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.*;
 import net.minecraft.core.particles.ParticleOptions;
-import net.minecraft.core.registries.Registries;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.util.AbortableIterationConsumer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.flag.FeatureFlagSet;
-import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.item.crafting.RecipeManager;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LightLayer;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.biome.BiomeManager;
-import net.minecraft.world.level.biome.Biomes;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntity;
@@ -45,10 +42,11 @@ import net.minecraft.world.level.storage.WritableLevelData;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.scores.Scoreboard;
-import net.minecraft.world.ticks.BlackholeTickAccess;
 import net.minecraft.world.ticks.LevelTickAccess;
 
 import javax.annotation.Nonnull;
+
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.lang.ref.WeakReference;
@@ -69,7 +67,7 @@ public class DummyWorld extends Level {
 
     protected DummyChunkSource chunkProvider = new DummyChunkSource(this);
     private final BiomeManager biomeManager;
-    protected WeakReference<Level> level;
+    public WeakReference<Level> level;
     protected final LevelLightEngine lighter;
     private final BlockPos.MutableBlockPos scratch = new BlockPos.MutableBlockPos();
     @Getter
@@ -82,6 +80,17 @@ public class DummyWorld extends Level {
         this.level = new WeakReference<>(level);
         this.lighter = new LevelLightEngine(chunkProvider, true, false);
         this.biomeManager = new BiomeManager(this, 0);
+    }
+
+    @NotNull
+    public Level getLevel() {
+        Level level = this.level.get();
+        if (level == null) {
+            level = Minecraft.getInstance().level;
+            this.level = new WeakReference<>(level);
+        }
+        assert level != null;
+        return level;
     }
 
     @Override
@@ -167,20 +176,12 @@ public class DummyWorld extends Level {
 
     @Override
     public Holder<Biome> getUncachedNoiseBiome(int pX, int pY, int pZ) {
-        Level level = this.level.get();
-        if (level != null) {
-            return level.getUncachedNoiseBiome(pX, pY, pZ);
-        }
-        return this.registryAccess().registryOrThrow(Registries.BIOME).getHolderOrThrow(Biomes.PLAINS);
+        return getLevel().getUncachedNoiseBiome(pX, pY, pZ);
     }
 
     @Override
     public Holder<Biome> getNoiseBiome(int pX, int pY, int pZ) {
-        Level level = this.level.get();
-        if (level != null) {
-            return level.getNoiseBiome(pX, pY, pZ);
-        }
-        return this.registryAccess().registryOrThrow(Registries.BIOME).getHolderOrThrow(Biomes.PLAINS);
+        return getLevel().getNoiseBiome(pX, pY, pZ);
     }
 
     @Override
@@ -190,65 +191,37 @@ public class DummyWorld extends Level {
 
     @Override
     public RegistryAccess registryAccess() {
-        Level level = this.level.get();
-        if (level != null) {
-            return level.registryAccess();
-        }
-        return Platform.getFrozenRegistry();
+        return getLevel().registryAccess();
     }
 
     @Override
     public FeatureFlagSet enabledFeatures() {
-        Level level = this.level.get();
-        if (level != null) {
-            return level.enabledFeatures();
-        }
-        return FeatureFlags.REGISTRY.allFlags();
+        return getLevel().enabledFeatures();
     }
 
     @Override
     public LevelTickAccess<Block> getBlockTicks() {
-        Level level = this.level.get();
-        if (level != null) {
-            return level.getBlockTicks();
-        }
-        return BlackholeTickAccess.emptyLevelList();
+        return getLevel().getBlockTicks();
     }
 
     @Override
     public LevelTickAccess<Fluid> getFluidTicks() {
-        Level level = this.level.get();
-        if (level != null) {
-            return level.getFluidTicks();
-        }
-        return BlackholeTickAccess.emptyLevelList();
+        return getLevel().getFluidTicks();
     }
 
     @Override
     public RecipeManager getRecipeManager() {
-        Level level = this.level.get();
-        if (level != null) {
-            level.getRecipeManager();
-        }
-        return Platform.isClient() ? Minecraft.getInstance().getConnection().getRecipeManager() : Platform.getMinecraftServer().getRecipeManager();
+        return getLevel().getRecipeManager();
     }
 
     @Override
     public int getFreeMapId() {
-        Level level = this.level.get();
-        if (level != null) {
-            return level.getFreeMapId();
-        }
-        return 0;
+        return getLevel().getFreeMapId();
     }
 
     @Override
     public Scoreboard getScoreboard() {
-        Level level = this.level.get();
-        if (level != null) {
-            return level.getScoreboard();
-        }
-        return new Scoreboard();
+        return getLevel().getScoreboard();
     }
 
     @Override
@@ -384,18 +357,8 @@ public class DummyWorld extends Level {
     }
 
     @Environment(EnvType.CLIENT)
+    @Getter @Setter
     private ParticleManager particleManager;
-
-    @Environment(EnvType.CLIENT)
-    public void setParticleManager(ParticleManager particleManager) {
-        this.particleManager = particleManager;
-    }
-
-    @Environment(EnvType.CLIENT)
-    @Nullable
-    public ParticleManager getParticleManager() {
-        return particleManager;
-    }
 
     public BlockState getBlockState(int x, int y, int z) {
         return getBlockState(scratch.set(x, y, z));