+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # ChangeLog
 
-* fix crash that happened when no datapack registries were loaded yet but regular ones needed to be accessed
+* fix TrackedDummyWorld leaking client level instances of ~300mb